pyproject.toml
[tool.poetry]
name = "pyannote-tests"
version = "0.1.0"
description = ""
authors = ["Kevin Diem <kg.diem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyannote-audio = "^3.3.1"
onnxruntime = "^1.18.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
